requirements.txt
# Dependencias de Flask (Backend)
Flask
Flask-CORS
gunicorn
python-dotenv

# Dependencias de Lógica (IA y PDF)
pdfplumber
pandas
PyMuPDF
google-genai==1.49.0
reportlab